backup button
Changed how the backup button works, it now tells the user that it's finished backing up the files. It also checks every install location for files to see if you even have any, and if not, warns that you may not have exported them in-game. Also added button that links to the RSI wiki on how to export controls